services/gateway-iot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
paho-mqtt==2.1.0
influxdb-client==1.48.0
psycopg2-binary==2.9.10
prometheus-client==0.21.1
python-dotenv==1.0.1

pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
